refactor: break out event restrictions module
