+++ rust/capture/src/event_restrictions/mod.rs
@@ -1,0 +1,14 @@
+mod manager;
+mod repository;
+mod types;
+
+// Re-export public API
+pub use manager::{EventRestrictionService, RestrictionManager};
+pub use repository::{EventRestrictionsRepository, RedisRestrictionsRepository, RestrictionEntry};
+pub use types::{
+    AppliedRestrictions, EventContext, Restriction, RestrictionFilters, RestrictionScope,
+    RestrictionType,
+};
+
+#[cfg(test)]
+pub use repository::testing::MockRestrictionsRepository;
